0.7.dev1 handle python 3 integer dividal returning float

+++ wwpdb/apps/workmanager/task_access/CifChecker.py
@@ -38,7 +38,7 @@ class CifChecker(BaseClass):
     def run(self):
         """
         """
-        numProc = multiprocessing.cpu_count() / 2
+        numProc = int(multiprocessing.cpu_count() / 2)
         mpu = MultiProcUtil(verbose = True)
         mpu.set(workerObj = self, workerMethod = "runMulti")
         mpu.setWorkingDir(self._sessionPath)
